api/requirements.txt
cycler==0.10.0
Django==1.11.5
django-json-widget==0.1.0
django-pdfkit==0.3.1
djangorestframework==3.6.4
html5lib==0.999999999
lxml==4.1.0
matplotlib==2.0.2
mimesis==0.0.10
numpy==1.13.3
olefile==0.44
pandas==0.20.3
pdfkit==0.6.0
Pillow==4.3.0
pip-autoremove==0.9.0
psycopg2==2.7.3.1
Pygments==2.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
python-dateutil==2.6.1
python-docx==0.8.6
pytz==2017.2
reportlab==3.4.0
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0
sklearn==0.0
tornado==5.0.1
transliterate==1.10
webencodings==0.5.1
wkhtmltopdf==0.2
xhtml2pdf==0.0.6
XlsxWriter==1.0.0
